+++ src/grpc_interceptor/__init__.py
@@ -1,7 +1,7 @@
 """Simplified Python gRPC interceptors."""
 
 from grpc_interceptor.client import ClientCallDetails, ClientInterceptor, AsyncClientInterceptor
-from grpc_interceptor.aio_channel import insecure_channel, secure_channel
+from grpc_interceptor._aio_channel import insecure_channel, secure_channel
 from grpc_interceptor.exception_to_status import (
     AsyncExceptionToStatusInterceptor,
     ExceptionToStatusInterceptor,
